Moved GroundCheck and CameraScript into namespaces
