--- Assets/Scripts/GrapplingHook.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 5b34195b1a8e8c743a4f805ff384e6ee
