--- Assets/Scripts/GrapplingHookTip.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 1d0ff98e7a91b8c4dbd65bd34a650a35
