--- Assets/Scripts/PlayerController.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: cfc7562e4642d724db06d4c63b0b450c
