+++ Assets/Scripts/TimeKnight/Core/GroundCheck.cs
@@ -1,0 +1,25 @@
+using UnityEngine;
+
+namespace TimeKnight.Core
+{
+    public class GroundCheck : MonoBehaviour
+    {
+        public bool IsGrounded { get; private set; }
+
+        private void OnTriggerEnter2D(Collider2D collision)
+        {
+            if ("Ground" == LayerMask.LayerToName(collision.gameObject.layer))
+            {
+                IsGrounded = true;
+            }
+        }
+
+        private void OnTriggerExit2D(Collider2D collision)
+        {
+            if ("Ground" == LayerMask.LayerToName(collision.gameObject.layer))
+            {
+                IsGrounded = false;
+            }
+        }
+    }
+}
