+++ Assets/Scripts/TimeKnight/Core/Interaction/PlayerInteraction.cs
@@ -1,4 +1,3 @@
-using System;
 using UnityEngine;
 
 namespace TimeKnight.Core.Interaction
